corporate_agent/requirements.txt
gradio
python-docx
docx2txt
sentence-transformers
faiss-cpu
openai
python-dotenv